Create SkipLoopPlus
This function takes a start number, skip number, and end number. These numbers are used to create loops that either increment up or down depending on the start number and end number. They then skip over the skip number in either loop.

The original problem asked to create a loop which increments up and skips the number 5. The function above is more robust.